move base tests to under tests folder
