+++ base/src/numbers.rs
@@ -189,27 +189,3 @@ impl From<BigInt256> for ark_ff::BigInteger256 {
         Self::read(&i.0[..]).unwrap()
     }
 }
-
-#[cfg(test)]
-pub mod tests {
-    use crate::numbers::Amount;
-    use crate::numbers::BigInt256;
-
-    #[test]
-    pub fn test_amount_to_formatted_string() {
-        assert_eq!(Amount(0).to_formatted_string(), "0.000000000");
-        assert_eq!(Amount(3).to_formatted_string(), "0.000000003");
-        assert_eq!(Amount(1000000003).to_formatted_string(), "1.000000003");
-        assert_eq!(Amount(1000000030).to_formatted_string(), "1.000000030");
-        assert_eq!(Amount(1300000000).to_formatted_string(), "1.300000000");
-        assert_eq!(Amount(1000000000).to_formatted_string(), "1.000000000");
-    }
-
-    #[test]
-    fn test_convert_bigint_to_arkworks_zero() {
-        use ark_ff::BigInteger256;
-        let i = BigInt256([0; 32]);
-        let ark_i: BigInteger256 = i.into();
-        assert_eq!(ark_i, BigInteger256::default())
-    }
-}
